Move DownloadManager to a separated feature module
