+++ core/common/src/main/java/com/youtubemusic/core/common/PlaylistsAdapter.kt
@@ -1,4 +1,4 @@
-package com.yurii.youtubemusic.screens.youtube.playlists
+package com.youtubemusic.core.common
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,9 +6,8 @@ import androidx.core.content.ContextCompat
 import androidx.paging.PagingDataAdapter
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import com.youtubemusic.core.common.databinding.ItemPlaylistBinding
 import com.youtubemusic.core.model.YouTubePlaylist
-import com.yurii.youtubemusic.R
-import com.yurii.youtubemusic.databinding.ItemPlaylistBinding
 
 class PlaylistsAdapter(private val onSelectedPlaylist: (YouTubePlaylist) -> Unit) :
     PagingDataAdapter<YouTubePlaylist, PlaylistsAdapter.PlaylistViewHolder>(Comparator) {
@@ -34,7 +33,7 @@ class PlaylistsAdapter(private val onSelectedPlaylist: (YouTubePlaylist) -> Unit
     inner class PlaylistViewHolder(val binding: ItemPlaylistBinding) : RecyclerView.ViewHolder(binding.root) {
         fun bind(playlist: YouTubePlaylist) {
             val backgroundColor = ContextCompat.getColor(
-                binding.root.context, if (selectedPlaylist == playlist) R.color.lightGray else R.color.white
+                binding.root.context, if (selectedPlaylist == playlist) android.R.color.darker_gray else android.R.color.white
             )
             binding.apply {
                 this.playlist = playlist
