+++ core/common/src/main/java/com/youtubemusic/core/common/XmlAttributes.kt
@@ -1,4 +1,4 @@
-package com.yurii.youtubemusic.utilities
+package com.youtubemusic.core.common
 
 import android.view.View
 import android.widget.ImageView
@@ -7,8 +7,6 @@ import androidx.core.view.isVisible
 import androidx.databinding.BindingAdapter
 import coil.load
 import coil.transform.RoundedCornersTransformation
-import com.youtubemusic.core.common.setAnimatedText
-import com.yurii.youtubemusic.R
 import java.io.File
