+++ core/common/src/main/res/drawable/ic_loading_image_error.xml
