+++ core/common/src/main/res/layout/dialog_error.xml
