+++ core/common/src/main/res/layout/item_loading.xml
