+++ core/common/src/main/res/layout/item_playlist.xml
@@ -31,7 +31,6 @@
                 android:layout_width="70dp"
                 android:layout_height="50dp"
                 android:layout_marginEnd="5dp"
-                android:contentDescription="@string/thumbnails_of_playlist"
                 app:imageUrl="@{playlist.thumbnailUrl}"
                 tools:src="@tools:sample/backgrounds/scenic" />
