+++ core/common/src/main/res/values/strings.xml
@@ -11,4 +11,5 @@
     <string name="dialog_cancel">Cancel</string>
     <string name="dialog_confirm_deletion_playlist_title">Delete playlist</string>
     <string name="dialog_confirm_deletion_playlist_message">Are you sure you want to delete this playlist?</string>
+    <string name="label_videos_count">%d videos</string>
 </resources>
