+++ feature/download-manager/build.gradle
@@ -1,11 +1,13 @@
 plugins {
     id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
+    id 'kotlin-android'
     id 'kotlin-kapt'
+    id 'kotlin-parcelize'
+    id 'com.google.dagger.hilt.android'
 }
 
 android {
-    namespace 'com.youtubemusic.core.common'
+    namespace 'com.youtubemusic.feature.download_manager'
     compileSdk 33
 
     defaultConfig {
@@ -29,20 +31,28 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+
     dataBinding {
         enabled = true
     }
 }
 
 dependencies {
-
+    implementation project(':core:common')
+    implementation project(':core:data')
+    implementation project(':core:model')
+    implementation project(':core:downloader:youtube')
+    implementation "androidx.fragment:fragment-ktx:$fragment_ktx_version"
+    implementation "com.github.shuhart:stepview:$stepview_version"
+    implementation "com.google.dagger:hilt-android:$hilt_version"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
+    kapt "com.google.dagger:hilt-compiler:$hilt_version"
+    implementation "androidx.fragment:fragment-ktx:$fragment_ktx_version"
+    implementation "androidx.paging:paging-runtime-ktx:$paging_version"
+    implementation "com.github.yogacp:android-viewbinding:$view_biding_delegate_version"
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.8.0'
-    implementation "org.threeten:threetenbp:$threetenbp_version"
-    implementation "io.coil-kt:coil:$coil_version"
-    implementation "androidx.paging:paging-runtime-ktx:$paging_version"
-    implementation project(path: ':core:model')
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
