+++ feature/download-manager/src/main/res/layout/dialog_add_youtube_playlist_sync.xml
@@ -33,16 +33,16 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/title"
             app:sv_animationType="Line"
-            app:sv_doneCircleColor="@color/colorAccent"
+            app:sv_doneCircleColor="?attr/colorAccent"
             app:sv_doneCircleRadius="12dp"
-            app:sv_doneStepLineColor="@color/colorAccent"
-            app:sv_doneStepMarkColor="@color/white"
-            app:sv_nextStepLineColor="@color/colorAccent"
-            app:sv_nextTextColor="@color/colorAccent"
-            app:sv_selectedCircleColor="@color/colorAccent"
+            app:sv_doneStepLineColor="?attr/colorAccent"
+            app:sv_doneStepMarkColor="@android:color/white"
+            app:sv_nextStepLineColor="?attr/colorAccent"
+            app:sv_nextTextColor="?attr/colorAccent"
+            app:sv_selectedCircleColor="?attr/colorAccent"
             app:sv_selectedCircleRadius="12dp"
-            app:sv_selectedStepNumberColor="@color/white"
-            app:sv_selectedTextColor="@color/colorAccent"
+            app:sv_selectedStepNumberColor="@android:color/white"
+            app:sv_selectedTextColor="?attr/colorAccent"
             app:sv_stepLineWidth="1dp"
             app:sv_stepPadding="4dp"
             app:sv_stepViewStyle="@style/StepView"
