+++ feature/download-manager/src/main/res/layout/item_headline.xml
@@ -5,7 +5,7 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@color/lightGray"
+        android:background="@android:color/darker_gray"
         android:gravity="center_vertical"
         android:orientation="horizontal">
