+++ feature/download-manager/src/main/res/values/array.xml
