+++ feature/download-manager/src/main/res/values/strings.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="label_add">Add</string>
+    <string name="label_download_manager">Download Manager</string>
+    <string name="label_playlists">Playlists</string>
+    <string name="label_ok">OK</string>
+    <string name="label_cancel">Cancel</string>
+    <string name="label_no_error_message">No error message</string>
+    <string name="label_playlist_sync_binds">Playlist sync binds</string>
+    <string name="label_downloading">Downloading</string>
+    <string name="label_cancel_all">Cancel all</string>
+    <string name="label_size_progress">%1$.1f / %2$.1f M</string>
+    <string name="hint_select_you_tube_playlist">Select YouTube playlist from which you want to download music automatically</string>
+    <string name="label_not_found_play_lists">It seems you haven\'t yet created a playlist. Please create a playlist and add your first video</string>
+    <string name="label_next">Next</string>
+    <string name="label_back">Back</string>
+    <string name="label_youtube_playlists_synchronization">YouTube Playlists synchronization</string>
+    <string name="hint_all_playlists_synchronized">You\'ve synchronized all your playlists</string>
+    <string name="label_no_playlist">You haven\'t created any playlist yet</string>
+    <string name="label_error">Error: %s</string>
+    <string name="hint_select_app_playlists">Now select the application playlists to which downloaded media items should be assigned. Note that by default all are assigned to All playlist</string>
+    <string name="label_id">ID: %s</string>
+</resources>
