+++ settings.gradle
@@ -24,3 +24,4 @@ include ':core:downloader:youtube'
 include ':core:player'
 include ':feature:equalizer'
 include ':feature:playlist-editor'
+include ':feature:download-manager'
